+++ changelog.md
@@ -1,4 +1,11 @@
 # Changelog
 
+**./src/generation/person-int.js**
+* 'chooseRandomInterviewDay' can now be called publicly as 'chooseInterviewDay'
+
+---
+
 **./src/generate-database-entries.js - coordinateGeneration**
-* Added 'personInt.chooseID' call for 'referralSourceID'
+* Added calls from 'personInt'
+	* chooseWageSubsidyFlag
+	* chooseInterviewDay